+++ mopidy/backends/dummy.py
@@ -39,18 +39,16 @@ class DummyLibraryProvider(base.BaseLibraryProvider):
         self.dummy_library = []
 
     def find_exact(self, **query):
-        return Playlist()
+        return []
 
     def lookup(self, uri):
-        matches = filter(lambda t: uri == t.uri, self.dummy_library)
-        if matches:
-            return matches[0]
+        return filter(lambda t: uri == t.uri, self.dummy_library)
 
     def refresh(self, uri=None):
         pass
 
     def search(self, **query):
-        return Playlist()
+        return []
 
 
 class DummyPlaybackProvider(base.BasePlaybackProvider):