+++ mopidy/core/tracklist.py
@@ -23,7 +23,7 @@ class TracklistController(object):
     @property
     def tl_tracks(self):
         """
-        List of two-tuples of (TLID integer, :class:`mopidy.models.Track`).
+        List of :class:`mopidy.models.TlTrack`.
 
         Read-only.
         """
@@ -72,8 +72,7 @@ class TracklistController(object):
         :type at_position: int or :class:`None`
         :param increase_version: if the tracklist version should be increased
         :type increase_version: :class:`True` or :class:`False`
-        :rtype: two-tuple of (TLID integer, :class:`mopidy.models.Track`) that
-            was added to the tracklist
+        :rtype: :class:`mopidy.models.TlTrack` that was added to the tracklist
         """
         assert at_position <= len(self._tl_tracks), \
             'at_position can not be greater than tracklist length'
@@ -117,22 +116,20 @@ class TracklistController(object):
         self._tl_tracks = []
         self.version += 1
 
-    def get(self, **criteria):
+    def filter(self, **criteria):
         """
-        Get track by given criterias from tracklist.
-
-        Raises :exc:`LookupError` if a unique match is not found.
+        Filter the tracklist by the given criterias.
 
         Examples::
 
-            get(tlid=7)             # Returns track with TLID 7 (tracklist ID)
-            get(id=1)               # Returns track with ID 1
-            get(uri='xyz')          # Returns track with URI 'xyz'
-            get(id=1, uri='xyz')    # Returns track with ID 1 and URI 'xyz'
+            filter(tlid=7)           # Returns track with TLID 7 (tracklist ID)
+            filter(id=1)             # Returns track with ID 1
+            filter(uri='xyz')        # Returns track with URI 'xyz'
+            filter(id=1, uri='xyz')  # Returns track with ID 1 and URI 'xyz'
 
         :param criteria: on or more criteria to match by
         :type criteria: dict
-        :rtype: two-tuple (TLID integer, :class:`mopidy.models.Track`)
+        :rtype: list of :class:`mopidy.models.TlTrack`
         """
         matches = self._tl_tracks
         for (key, value) in criteria.iteritems():
@@ -141,24 +138,16 @@ class TracklistController(object):
             else:
                 matches = filter(
                     lambda ct: getattr(ct.track, key) == value, matches)
-        if len(matches) == 1:
-            return matches[0]
-        criteria_string = ', '.join(
-            ['%s=%s' % (k, v) for (k, v) in criteria.iteritems()])
-        if len(matches) == 0:
-            raise LookupError('"%s" match no tracks' % criteria_string)
-        else:
-            raise LookupError('"%s" match multiple tracks' % criteria_string)
+        return matches
 
     def index(self, tl_track):
         """
-        Get index of the given (TLID integer, :class:`mopidy.models.Track`)
-        two-tuple in the tracklist.
+        Get index of the given :class:`mopidy.models.TlTrack` in the tracklist.
 
         Raises :exc:`ValueError` if not found.
 
         :param tl_track: track to find the index of
-        :type tl_track: two-tuple (TLID integer, :class:`mopidy.models.Track`)
+        :type tl_track: :class:`mopidy.models.TlTrack`
         :rtype: int
         """
         return self._tl_tracks.index(tl_track)
@@ -199,20 +188,23 @@ class TracklistController(object):
 
     def remove(self, **criteria):
         """
-        Remove the track from the tracklist.
+        Remove the matching tracks from the tracklist.
 
-        Uses :meth:`get()` to lookup the track to remove.
+        Uses :meth:`filter()` to lookup the tracks to remove.
 
         Triggers the :method:`mopidy.core.CoreListener.tracklist_changed`
         event.
 
         :param criteria: on or more criteria to match by
         :type criteria: dict
+        :rtype: list of :class:`mopidy.models.TlTrack` that was removed
         """
-        tl_track = self.get(**criteria)
-        position = self._tl_tracks.index(tl_track)
-        del self._tl_tracks[position]
+        tl_tracks = self.filter(**criteria)
+        for tl_track in tl_tracks:
+            position = self._tl_tracks.index(tl_track)
+            del self._tl_tracks[position]
         self.version += 1
+        return tl_tracks
 
     def shuffle(self, start=None, end=None):
         """
@@ -255,7 +247,7 @@ class TracklistController(object):
         :type start: int
         :param end: position after last track to include in slice
         :type end: int
-        :rtype: two-tuple of (TLID integer, :class:`mopidy.models.Track`)
+        :rtype: :class:`mopidy.models.TlTrack`
         """
         return self._tl_tracks[start:end]