+++ mopidy/models.py
@@ -14,7 +14,7 @@ class ImmutableObject(object):
 
     def __init__(self, *args, **kwargs):
         for key, value in kwargs.items():
-            if not hasattr(self, key):
+            if not hasattr(self, key) or callable(getattr(self, key)):
                 raise TypeError(
                     '__init__() got an unexpected keyword argument "%s"' %
                     key)
@@ -80,7 +80,7 @@ class ImmutableObject(object):
 
     def serialize(self):
         data = {}
-        data['__type__'] = self.__class__.__name__
+        data['__model__'] = self.__class__.__name__
         for key in self.__dict__.keys():
             public_key = key.lstrip('_')
             value = self.__dict__[key]
@@ -101,7 +101,7 @@ class ModelJSONEncoder(json.JSONEncoder):
 
         >>> import json
         >>> json.dumps({'a_track': Track(name='name')}, cls=ModelJSONEncoder)
-        '{"a_track": {"__type__": "Track", "name": "name"}}'
+        '{"a_track": {"__model__": "Track", "name": "name"}}'
 
     """
     def default(self, obj):
@@ -118,23 +118,16 @@ def model_json_decoder(dct):
 
         >>> import json
         >>> json.loads(
-        ...     '{"a_track": {"__type__": "Track", "name": "name"}}',
+        ...     '{"a_track": {"__model__": "Track", "name": "name"}}',
         ...     object_hook=model_json_decoder)
         {u'a_track': Track(artists=[], name=u'name')}
 
     """
-    if '__type__' in dct:
-        obj_type = dct.pop('__type__')
-        if obj_type == 'Album':
-            return Album(**dct)
-        if obj_type == 'Artist':
-            return Artist(**dct)
-        if obj_type == 'Playlist':
-            return Playlist(**dct)
-        if obj_type == 'TlTrack':
-            return TlTrack(**dct)
-        if obj_type == 'Track':
-            return Track(**dct)
+    if '__model__' in dct:
+        model_name = dct.pop('__model__')
+        cls = globals().get(model_name, None)
+        if issubclass(cls, ImmutableObject):
+            return cls(**dct)
     return dct