+++ tests/backends/base/library.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 import pykka
 
 from mopidy import core
-from mopidy.models import Playlist, Track, Album, Artist
+from mopidy.models import Track, Album, Artist
 
 from tests import unittest, path_to_data_dir
 
@@ -43,52 +43,52 @@ class LibraryControllerTest(object):
         pass
 
     def test_lookup(self):
-        track = self.library.lookup(self.tracks[0].uri)
-        self.assertEqual(track, self.tracks[0])
+        tracks = self.library.lookup(self.tracks[0].uri)
+        self.assertEqual(tracks, self.tracks[0:1])
 
     def test_lookup_unknown_track(self):
-        track = self.library.lookup('fake uri')
-        self.assertEquals(track, None)
+        tracks = self.library.lookup('fake uri')
+        self.assertEqual(tracks, [])
 
     def test_find_exact_no_hits(self):
         result = self.library.find_exact(track=['unknown track'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.find_exact(artist=['unknown artist'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.find_exact(album=['unknown artist'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
     def test_find_exact_artist(self):
         result = self.library.find_exact(artist=['artist1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.find_exact(artist=['artist2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_find_exact_track(self):
         result = self.library.find_exact(track=['track1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.find_exact(track=['track2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_find_exact_album(self):
         result = self.library.find_exact(album=['album1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.find_exact(album=['album2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_find_exact_uri(self):
         track_1_uri = 'file://' + path_to_data_dir('uri1')
         result = self.library.find_exact(uri=track_1_uri)
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         track_2_uri = 'file://' + path_to_data_dir('uri2')
         result = self.library.find_exact(uri=track_2_uri)
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_find_exact_wrong_type(self):
         test = lambda: self.library.find_exact(wrong=['test'])
@@ -106,57 +106,57 @@ class LibraryControllerTest(object):
 
     def test_search_no_hits(self):
         result = self.library.search(track=['unknown track'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.search(artist=['unknown artist'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.search(album=['unknown artist'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.search(uri=['unknown'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
         result = self.library.search(any=['unknown'])
-        self.assertEqual(result, Playlist())
+        self.assertEqual(result, [])
 
     def test_search_artist(self):
         result = self.library.search(artist=['Tist1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.search(artist=['Tist2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_search_track(self):
         result = self.library.search(track=['Rack1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.search(track=['Rack2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_search_album(self):
         result = self.library.search(album=['Bum1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.search(album=['Bum2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_search_uri(self):
         result = self.library.search(uri=['RI1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
         result = self.library.search(uri=['RI2'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[1:2]))
+        self.assertEqual(result, self.tracks[1:2])
 
     def test_search_any(self):
         result = self.library.search(any=['Tist1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
         result = self.library.search(any=['Rack1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
         result = self.library.search(any=['Bum1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
         result = self.library.search(any=['RI1'])
-        self.assertEqual(result, Playlist(tracks=self.tracks[:1]))
+        self.assertEqual(result, self.tracks[:1])
 
     def test_search_wrong_type(self):
         test = lambda: self.library.search(wrong=['test'])