+++ tests/models_test.py
@@ -79,6 +79,13 @@ class ArtistTest(unittest.TestCase):
         test = lambda: Artist(foo='baz')
         self.assertRaises(TypeError, test)
 
+    def test_invalid_kwarg_with_name_matching_method(self):
+        test = lambda: Artist(copy='baz')
+        self.assertRaises(TypeError, test)
+
+        test = lambda: Artist(serialize='baz')
+        self.assertRaises(TypeError, test)
+
     def test_repr(self):
         self.assertEquals(
             "Artist(name=u'name', uri=u'uri')",
@@ -86,14 +93,35 @@ class ArtistTest(unittest.TestCase):
 
     def test_serialize(self):
         self.assertDictEqual(
-            {'__type__': 'Artist', 'uri': 'uri', 'name': 'name'},
+            {'__model__': 'Artist', 'uri': 'uri', 'name': 'name'},
             Artist(uri='uri', name='name').serialize())
 
-    def test_to_json_and_Back(self):
+    def test_to_json_and_back(self):
         artist1 = Artist(uri='uri', name='name')
         serialized = json.dumps(artist1, cls=ModelJSONEncoder)
         artist2 = json.loads(serialized, object_hook=model_json_decoder)
         self.assertEqual(artist1, artist2)
+
+    def test_to_json_and_back_with_unknown_field(self):
+        artist = Artist(uri='uri', name='name').serialize()
+        artist['foo'] = 'foo'
+        serialized = json.dumps(artist)
+        test = lambda: json.loads(serialized, object_hook=model_json_decoder)
+        self.assertRaises(TypeError, test)
+
+    def test_to_json_and_back_with_field_matching_method(self):
+        artist = Artist(uri='uri', name='name').serialize()
+        artist['copy'] = 'foo'
+        serialized = json.dumps(artist)
+        test = lambda: json.loads(serialized, object_hook=model_json_decoder)
+        self.assertRaises(TypeError, test)
+
+    def test_to_json_and_back_with_field_matching_internal_field(self):
+        artist = Artist(uri='uri', name='name').serialize()
+        artist['__mro__'] = 'foo'
+        serialized = json.dumps(artist)
+        test = lambda: json.loads(serialized, object_hook=model_json_decoder)
+        self.assertRaises(TypeError, test)
 
     def test_eq_name(self):
         artist1 = Artist(name='name')
@@ -204,14 +232,14 @@ class AlbumTest(unittest.TestCase):
 
     def test_serialize_without_artists(self):
         self.assertDictEqual(
-            {'__type__': 'Album', 'uri': 'uri', 'name': 'name'},
+            {'__model__': 'Album', 'uri': 'uri', 'name': 'name'},
             Album(uri='uri', name='name').serialize())
 
     def test_serialize_with_artists(self):
         artist = Artist(name='foo')
         self.assertDictEqual(
-            {'__type__': 'Album', 'uri': 'uri', 'name': 'name', 'artists':
-                [artist.serialize()]},
+            {'__model__': 'Album', 'uri': 'uri', 'name': 'name',
+                'artists': [artist.serialize()]},
             Album(uri='uri', name='name', artists=[artist]).serialize())
 
     def test_to_json_and_back(self):
@@ -402,20 +430,20 @@ class TrackTest(unittest.TestCase):
 
     def test_serialize_without_artists(self):
         self.assertDictEqual(
-            {'__type__': 'Track', 'uri': 'uri', 'name': 'name'},
+            {'__model__': 'Track', 'uri': 'uri', 'name': 'name'},
             Track(uri='uri', name='name').serialize())
 
     def test_serialize_with_artists(self):
         artist = Artist(name='foo')
         self.assertDictEqual(
-            {'__type__': 'Track', 'uri': 'uri', 'name': 'name',
+            {'__model__': 'Track', 'uri': 'uri', 'name': 'name',
                 'artists': [artist.serialize()]},
             Track(uri='uri', name='name', artists=[artist]).serialize())
 
     def test_serialize_with_album(self):
         album = Album(name='foo')
         self.assertDictEqual(
-            {'__type__': 'Track', 'uri': 'uri', 'name': 'name',
+            {'__model__': 'Track', 'uri': 'uri', 'name': 'name',
                 'album': album.serialize()},
             Track(uri='uri', name='name', album=album).serialize())
 
@@ -618,7 +646,7 @@ class TlTrackTest(unittest.TestCase):
     def test_serialize(self):
         track = Track(uri='uri', name='name')
         self.assertDictEqual(
-            {'__type__': 'TlTrack', 'tlid': 123, 'track': track.serialize()},
+            {'__model__': 'TlTrack', 'tlid': 123, 'track': track.serialize()},
             TlTrack(tlid=123, track=track).serialize())
 
     def test_to_json_and_back(self):
@@ -752,13 +780,13 @@ class PlaylistTest(unittest.TestCase):
 
     def test_serialize_without_tracks(self):
         self.assertDictEqual(
-            {'__type__': 'Playlist', 'uri': 'uri', 'name': 'name'},
+            {'__model__': 'Playlist', 'uri': 'uri', 'name': 'name'},
             Playlist(uri='uri', name='name').serialize())
 
     def test_serialize_with_tracks(self):
         track = Track(name='foo')
         self.assertDictEqual(
-            {'__type__': 'Playlist', 'uri': 'uri', 'name': 'name',
+            {'__model__': 'Playlist', 'uri': 'uri', 'name': 'name',
                 'tracks': [track.serialize()]},
             Playlist(uri='uri', name='name', tracks=[track]).serialize())