+++ docs/changes.rst
@@ -128,6 +128,23 @@ backends:
   means that you now can select playlists to queue and play from the Ubuntu
   Sound Menu.
 
+- :meth:`mopidy.core.LibraryController.find_exact` and
+  :meth:`mopidy.core.LibraryController.search` now returns plain lists of
+  tracks instead of playlist objects.
+
+- :meth:`mopidy.core.TracklistController.get` has been replaced by
+  :meth:`mopidy.core.TracklistController.filter`.
+
+- :meth:`mopidy.core.PlaylistsController.get` has been replaced by
+  :meth:`mopidy.core.PlaylistsController.filter`.
+
+- :meth:`mopidy.core.TracklistController.remove` can now remove multiple
+  tracks, and returns the tracks it removed.
+
+- :meth:`mopidy.core.LibraryController.lookup` now returns a list of tracks.
+  This makes it possible to support lookup of artist or album URIs which then
+  can expand to a list of tracks.
+
 **Bug fixes**
 
 - :issue:`218`: The MPD commands ``listplaylist`` and ``listplaylistinfo`` now
@@ -142,6 +159,10 @@ backends:
 - :issue:`236`: The ``mopidy-scan`` command failed to include tags from ALAC
   files (Apple lossless) because it didn't support multiple tag messages from
   GStreamer per track it scanned.
+
+- :issue:`246`: The MPD command ``list album artist ""`` and similar
+  ``search``, ``find``, and ``list`` commands with empty filter values caused a
+  :exc:`LookupError`, but should have been ignored by the MPD server.
 
 
 v0.8.1 (2012-10-30)